ref(node): Allow node stack parser to work in browser context  (#5135)
