+++ packages/node/test/context-lines.test.ts
@@ -3,7 +3,7 @@ import * as fs from 'fs';
 
 import { parseStackFrames } from '../src/eventbuilder';
 import { ContextLines, resetFileContentCache } from '../src/integrations/contextlines';
-import { defaultStackParser } from '../src/stack-parser';
+import { defaultStackParser } from '../src/sdk';
 import { getError } from './helper/error';
 
 describe('ContextLines', () => {
